broadcast_tx: Fix smartbit pushtx usage, also reorder services
smartbit's API actually takes direct JSON, not the typical POST
format. Thus, smartbit was never used for broadcasts.

Also reorders API usage consistently to Bitpay first, Smartbit
second, and Blockchain last. Blockchain is notoriously
unreliable. I have seen it go down many times and be several
hours behind in others.

Today while testing, repeatedly it returned a 200 for pushtx and
yet the transaction was not pushed to the network. I pushed it
with smartbit and could even then find it on Blockchain, itself.
Sending a 200 for a transaction that was not pushed is completely
unacceptable from a programatic usage standpoint as no other
APIs will be used in turn. Hopefully, this reordering should
help minimize any Blockchain "oddities" as it will be the last
API in line.