Fixes #30: Calculate fee with only used unspents, warn about default fee (#32)
Before this commit there was an issue where combine=False send()'s
would calculate the fee with all unspents. Now it will recalculate
the fee as it goes along adding unspents to use the minimum number
of unspents and the smallest fee.

Note that is this is very unoptimized and some unspents add more
fees than they are worth.

Adds one test case that was not broken before this point and another
for validating that additional, unnecessary unspents do not raise
the fee.

Also introduce logging, add warnings if the fee API cannot be reached
and the default or cached is used.

The warnings will only be seen in some cases depending on the logging
level set, but are always caught with failing test cases and can
further aid in debugging.